console tools and LightFX optimization and bugfix

+++ alienfan-tools/alienfan-SDK/alienfan-SDK.h
@@ -17,7 +17,7 @@ namespace AlienFan_SDK {
 	};
 
 	struct ALIENFAN_COMMAND {
-		short com;
+		byte /*short*/ com;
 		byte sub;
 	};
